requirements.txt
fastapi
uvicorn
pillow
replicate
python-multipart
